requirements.txt
aiohttp
fake_useragent
pydantic
beautifulsoup4
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
